Re-organize new functions
